build(Spring-Boot): Upgrade to 2.4.13 #70
- Required upgrading jedis 3.1.0->3.7.0 and restarting Redis
- Required breaking websocket endpoints into their own controllers
