+++ src/main/java/org/wise/portal/spring/impl/WebSocketConfig.java
@@ -25,7 +25,7 @@ public class WebSocketConfig implements WebSocketMessageBrokerConfigurer {
 
   @Override
   public void registerStompEndpoints(StompEndpointRegistry registry) {
-    registry.addEndpoint("/websocket").setAllowedOrigins("*").withSockJS();
+    registry.addEndpoint("/websocket").setAllowedOriginPatterns("*").withSockJS();
   }
 
   @Override
